+++ packages/components-native/src/ActivityIndicator/ActivityIndicator.tsx
@@ -3,9 +3,7 @@ import type { ActivityIndicatorProps } from "react-native";
 import { ActivityIndicator } from "react-native";
 import { useAtlantisTheme } from "../AtlantisThemeContext";
 
-export function JobberActivityIndicator(
-  props: ActivityIndicatorProps,
-): JSX.Element {
+export function JobberActivityIndicator(props: ActivityIndicatorProps) {
   const { tokens } = useAtlantisTheme();
 
   return (