+++ packages/components-native/src/ErrorMessageWrapper/context/ErrorMessageProvider.tsx
@@ -13,9 +13,7 @@ interface ErrorMessageProviderProps {
   readonly children: ReactNode;
 }
 
-export function ErrorMessageProvider({
-  children,
-}: ErrorMessageProviderProps): JSX.Element {
+export function ErrorMessageProvider({ children }: ErrorMessageProviderProps) {
   const [elements, setElements] = useState<
     ErrorMessageContextProps["elements"]
   >({});
@@ -51,13 +49,13 @@ export function ErrorMessageProvider({
   }
 }
 
-function getMeasure(ref: RefObject<View>) {
+function getMeasure(ref: RefObject<View | null>) {
   return function measure(...args: Parameters<NativeMethods["measureLayout"]>) {
     ref.current?.measureLayout(...args);
   };
 }
 
-function getAccessibilityFocus(ref: RefObject<View>) {
+function getAccessibilityFocus(ref: RefObject<View | null>) {
   return function accessibilityFocus() {
     const reactTag = findNodeHandle(ref.current);
     reactTag &&