+++ packages/components-native/src/Menu/types.ts
@@ -1,4 +1,5 @@
 import type { IconColorNames, IconNames } from "@jobber/design";
+import type { ReactElement } from "react";
 import type { TextAlign } from "../Typography";
 
 export interface MenuOptionProps {
@@ -15,9 +16,14 @@ export interface MenuOptionInternalProps extends MenuOptionProps {
   setOpen: (bool: boolean) => void;
 }
 
+export interface CustomActivatorProps {
+  readonly onPress?: () => void;
+  readonly onLongPress?: () => void;
+}
+
 export interface MenuProps {
   readonly menuOptions?: MenuOptionProps[];
-  readonly customActivator?: JSX.Element;
+  readonly customActivator?: ReactElement<CustomActivatorProps>;
 }
 
 export interface OverlayProp {