+++ packages/components/src/Combobox/components/ComboboxActivator/ComboboxActivator.test.tsx
@@ -73,7 +73,7 @@ describe("ComboboxActivator", () => {
 });
 
 function renderComboboxActivator(
-  child: ReactElement | ((args: ComboboxCustomActivatorProps) => JSX.Element),
+  child: ReactElement | ((args: ComboboxCustomActivatorProps) => ReactElement),
   open: boolean,
 ) {
   return render(