+++ packages/components/src/DataList/components/DataListLayoutActions/DataListLayoutActions.test.tsx
@@ -1,4 +1,4 @@
-import type { PropsWithChildren } from "react";
+import type { PropsWithChildren, ReactElement } from "react";
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { DataListLayoutActions } from "./DataListLayoutActions";
@@ -11,7 +11,7 @@ import {
 } from "../../context/DataListLayoutContext";
 
 const mockSetHasInLayoutActions = jest.fn();
-const mockItemActionComponent = jest.fn<JSX.Element | undefined, []>(() => (
+const mockItemActionComponent = jest.fn<ReactElement | undefined, []>(() => (
   <DataListItemActions>
     <DataListAction label="Edit" />
     <DataListAction label="Email" />