+++ packages/components/src/FormField/hooks/useFormFieldWrapperStyles.ts
@@ -24,8 +24,8 @@ export interface useFormFieldWrapperStylesProps
    * In v2 components, this is typically undefined.
    */
   readonly max?: number | string;
-  suffixRef?: RefObject<HTMLDivElement>;
-  prefixRef?: RefObject<HTMLDivElement>;
+  suffixRef?: RefObject<HTMLDivElement | null>;
+  prefixRef?: RefObject<HTMLDivElement | null>;
   showMiniLabel?: boolean;
 }